kafka-client/requirements.txt
kafka-python==2.0.2
elasticsearch==8.8.0